+++ NEWS.md
@@ -1,5 +1,10 @@
 # QFeatures 1.1.0
 
+## QFeatures 1.1.4
+
+- Added `replaceRowDataCols` and `removeRowDataCols`, two functions 
+  to streamline manipulation of rowData within a `QFeature` object. 
+
 ## QFeatures 1.1.3
 
 - Added `countUniqueFeatures`, a function to count the number of